Add setup wizard
Added a "ZMK: Add keyboard" command which works like the user setup
script. After the user selects all options, it copies the keyboard files
into the repo and adds items to build.yaml.

This also scans the user's repo for custom boards and shields and
includes them in the keyboard list.